feat: add Dynamic class
This commit adds the Dynamic utility class, which provides an observable type that can be subscribed to locally. This allows any class with access to the dynamic variable to bind some closure to the variable in a local disposebag, which will run when the variable changes. The disposebag handles disposal of any bound closure when the class goes out of scope.